Begin work on compaction. Came up with nice naming scheme for streams
