+++ packages/punchcard/lib/firehose/event.ts
@@ -1,37 +1,3 @@
-export interface S3Event {
-  Records: Array<{
-    eventVersion: string;
-    eventSource: string;
-    awsRegion: string;
-    eventTime: string;
-    eventName: string;
-    requestParameters: {
-      sourceIPAddress: string;
-    };
-    responseElements: {
-      'x-amz-request-id': string;
-      'x-amz-id-2': string;
-    };
-    s3: {
-      s3SchemaVersion: string;
-      configurationId: string;
-      bucket: {
-        name: string;
-        ownerIdentity: {
-          principalId: string;
-        };
-        arn: string;
-      };
-      object: {
-        key: string;
-        size: number;
-        eTag: string;
-        sequencer: string;
-      };
-    };
-  }>;
-}
-
 export interface FirehoseEvent {
   records: Array<{
     recordId: string;
