+++ packages/punchcard/lib/kinesis/index.ts
@@ -1,3 +1,5 @@
+export * from './client';
 export * from './collector';
 export * from './event';
+export * from './records';
 export * from './stream';
