+++ packages/punchcard/lib/s3/event.ts
@@ -1,0 +1,33 @@
+export interface Event {
+  Records: Array<{
+    eventVersion: string;
+    eventSource: string;
+    awsRegion: string;
+    eventTime: string;
+    eventName: string;
+    requestParameters: {
+      sourceIPAddress: string;
+    };
+    responseElements: {
+      'x-amz-request-id': string;
+      'x-amz-id-2': string;
+    };
+    s3: {
+      s3SchemaVersion: string;
+      configurationId: string;
+      bucket: {
+        name: string;
+        ownerIdentity: {
+          principalId: string;
+        };
+        arn: string;
+      };
+      object: {
+        key: string;
+        size: number;
+        eTag: string;
+        sequencer: string;
+      };
+    };
+  }>;
+}
