+++ packages/punchcard/lib/s3/index.ts
@@ -1,1 +1,2 @@
 export * from './bucket';
+export * from './event';
