+++ packages/punchcard/lib/sqs/messages.ts
@@ -7,14 +7,18 @@ import { Queue } from './queue';
 
 export type Config = Stream.Config & events.SqsEventSourceProps;
 
-export class QueueStream<T, D extends any[]> extends Stream<Event, T, D, Config>  {
-  constructor(public readonly queue: Queue<any>, previous: QueueStream<any, any>, input: {
+/**
+ * A `Stream` of Messages from a SQS Queue.
+ */
+export class Messages<T, D extends any[]> extends Stream<Event, T, D, Config>  {
+  constructor(public readonly queue: Queue<any>, previous: Messages<any, any>, input: {
     depends: D;
     handle: (value: AsyncIterableIterator<any>, deps: Clients<D>) => AsyncIterableIterator<T>;
   }) {
     super(previous, input.handle, input.depends);
   }
 
+  // TODO: this should be passed in at instantiation time!!!
   public eventSource(props?: Config) {
     return new events.SqsEventSource(this.queue.resource, props);
   }
@@ -22,7 +26,7 @@ export class QueueStream<T, D extends any[]> extends Stream<Event, T, D, Config>
   public chain<U, D2 extends any[]>(input: {
     depends: D2;
     handle: (value: AsyncIterableIterator<T>, deps: Clients<D2>) => AsyncIterableIterator<U>;
-  }): QueueStream<U, D2> {
-    return new QueueStream<U, D2>(this.queue, this, input);
+  }): Messages<U, D2> {
+    return new Messages<U, D2>(this.queue, this, input);
   }
 }
